NetworkMessages.cs not only contains the NetworkMessageType enum
