--- Assets/Scripts/NetworkBehaviourTemplate.cs.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f27092929c0162d4397271242dffa865
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
