+++ Assets/Scripts/Networking/ServerObject.cs
@@ -12,8 +12,6 @@ public class ServerObject : MonoBehaviour
 
     void Start ()
     {
-        onServer = gameObject.layer == LayerMask.NameToLayer("Server");
-
         if (onServer)
         {
             // Check if the resource name is valid
@@ -38,5 +36,52 @@ public class ServerObject : MonoBehaviour
             GameServer.Instance.serverObjects.Remove(this);
             GameServer.Instance.SendMessageDestroyServerObject(this);
         }
+    }
+}
+
+[System.Serializable]
+struct MessageServerObject
+{
+    public float time;
+    public string name;
+    public string resourceName;
+    public bool hasParent;
+    public int parentInstanceID;
+    public int instanceID;
+    public Vector3 localPosition;
+    public Quaternion localRotation;
+    public Vector3 localScale;
+
+    public MessageServerObject(ServerObject serverObject)
+    {
+        if (serverObject.transform.parent != null)
+        {
+            parentInstanceID = serverObject.transform.parent.GetInstanceID();
+            hasParent = true;
+        }
+        else
+        {
+            parentInstanceID = 0;
+            hasParent = false;
+        }
+
+        time = serverObject.lastUpdate = Time.time;
+        name = serverObject.name;
+        resourceName = serverObject.resourceName;
+        instanceID = serverObject.transform.GetInstanceID();
+        localPosition = serverObject.transform.localPosition;
+        localRotation = serverObject.transform.localRotation;
+        localScale = serverObject.transform.localScale;
+    }
+}
+
+[System.Serializable]
+struct MessageDestroyServerObject
+{
+    public int instanceID;
+
+    public MessageDestroyServerObject(ServerObject serverObject)
+    {
+        instanceID = serverObject.transform.GetInstanceID();
     }
 }
